+++ agtern/common/api.py
@@ -3,8 +3,8 @@ Post-MVP: This file will provide functions that will query the AgTern web server
 
 from typing import List
 
-from ..backend import get_all_internships
-from ...models import Internship
+from .models import Internship
+from ..server import get_all_internships
 
 
 def api_get_all_internships() -> List[Internship]:
