--- agtern/pipelines/data.py
@@ -1,41 +1,0 @@
-from os import makedirs
-from os.path import isfile
-from pathlib import Path
-from typing import Union
-
-
-class DataFile:
-    def __init__(
-        self,
-        filename: str,
-        is_temp: bool = False,
-        default_data: Union[str, None] = None,
-        create_on_init: bool = True
-    ):
-        self.filename = filename
-        self.is_temp = is_temp
-        if is_temp:
-            self.path = Path.cwd().joinpath("data", "temp", filename)
-        else:
-            self.path = Path.cwd().joinpath("data", filename)
-        
-        if create_on_init:
-            assert self.create(default_data)
-
-    def exists(self):
-        return isfile(self.path)
-
-    def create(self, default_data: Union[str, None] = None):
-        """Creates a data file if it doesn't exist. Returns true if successful."""
-        if self.exists():
-            return True
-        try:
-            if self.is_temp:
-                makedirs(Path.cwd().joinpath("data", "temp"), exist_ok=True)
-            else:
-                makedirs(Path.cwd().joinpath("data"), exist_ok=True)
-            with open(self.path, "w") as f:
-                f.write(default_data if default_data else "")
-        except OSError:
-            pass
-        return self.exists()
