--- agtern/pipelines/frontend/__init__.py
@@ -1,1 +1,0 @@
-from .api import api_get_all_internships
