+++ agtern/server/scrapers/__init__.py
@@ -1,3 +1,3 @@
 from .extractor import update_internship
-from .scraper import start_scraper
+from .scraper import scrape, start_scraper
 from .server import get_all_internships, start_server
